refactor: route modal
