+++ src/app/home/components/route-modal/route-modal.component.scss
@@ -2,7 +2,7 @@
   .header {
     display: flex;
     align-items: center;
-    width: 500px;
+    
     margin-top: 0.5rem;
     margin-bottom: 0;
   }
@@ -12,13 +12,25 @@
   }
 
   .step-label {
-    width: 300px;
     display: flex;
     align-items: center;
+    width: 350px;
+
+    @media (max-width: 1200px) {
+      width: 300px;
+    }
+    @media (max-width: 768px) {
+      width: 200px;
+    }
   }
 
   .spacer {
     margin: 0 auto;
+  }
+
+  .dwell {
+    display: flex;
+    align-items: center;
   }
 
   .spinner-container {
