+++ src/app/home/components/route-modal/route-modal.component.ts
@@ -1,38 +1,16 @@
 import { StationStore } from '@admin/store/stations/stations.store';
-import { DatePipe, JsonPipe } from '@angular/common';
 import { Component, inject, OnInit, signal } from '@angular/core';
-import { MatButtonModule } from '@angular/material/button';
-import {
-  MAT_DIALOG_DATA,
-  MatDialogActions,
-  MatDialogClose,
-  MatDialogContent,
-  MatDialogRef,
-  MatDialogTitle,
-} from '@angular/material/dialog';
-import { MatDivider } from '@angular/material/divider';
-import { MatIcon } from '@angular/material/icon';
-import { MatProgressSpinner } from '@angular/material/progress-spinner';
-import { MatStepperModule } from '@angular/material/stepper';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { RouteModalData, RouteStop } from '@home/models/trip.models';
 import { RideStore } from '@shared/store/ride/ride.store';
+
+import { routeModalImports } from './route-modal.config';
+import { getDiffInMinutes } from './route-modal.utils';
 
 @Component({
   selector: 'app-route-modal',
   standalone: true,
-  imports: [
-    MatDialogTitle,
-    MatDialogContent,
-    MatDialogActions,
-    MatDialogClose,
-    MatButtonModule,
-    MatDivider,
-    MatIcon,
-    DatePipe,
-    MatProgressSpinner,
-    JsonPipe,
-    MatStepperModule,
-  ],
+  imports: routeModalImports,
   templateUrl: './route-modal.component.html',
   styleUrl: './route-modal.component.scss',
 })
@@ -45,54 +23,39 @@ export class RouteModalComponent implements OnInit {
 
   routeData = signal<RouteStop[]>([]);
 
-  rideStore = inject(RideStore);
+  private rideStore = inject(RideStore);
 
-  stationStore = inject(StationStore);
+  private stationStore = inject(StationStore);
 
-  ngOnInit(): void {
-    this.initRouteData();
-  }
-
-  async initRouteData() {
+  async ngOnInit() {
     this.isLoading.set(true);
     await this.rideStore.getRide(this.data.rideId);
     await this.stationStore.getStations();
 
+    this.initRouteData();
+    this.isLoading.set(false);
+  }
+
+  private initRouteData() {
     const ride = this.rideStore.ridesEntityMap()[this.data.rideId];
     const stationsMap = this.stationStore.stationsEntityMap();
-
     const stations = ride.path.map((stationId) => stationsMap[stationId]);
     const { segments } = ride.schedule;
 
     const routeStops: RouteStop[] = [];
-
     stations.forEach((station, idx) => {
-      const arrivalTime = idx > 0 ? segments[idx - 1].time[1] : '';
-      const departureTime = idx < segments.length ? segments[idx].time[0] : '';
+      const arrival = idx > 0 ? segments[idx - 1].time[1] : '';
+      const departure = idx < segments.length ? segments[idx].time[0] : '';
+      const dwellTime = getDiffInMinutes(arrival, departure);
 
       routeStops.push({
-        city: station.city,
-        arrival: arrivalTime,
-        departure: departureTime,
-        dwellTime: this.getTimeDifference(arrivalTime, departureTime),
+        station,
+        arrival,
+        departure,
+        dwellTime,
       });
     });
 
     this.routeData.set(routeStops);
-    this.isLoading.set(false);
-  }
-
-  private getTimeDifference(
-    isoTime1: string,
-    isoTime2: string,
-  ): number | undefined {
-    if (!isoTime1 || !isoTime2) {
-      return undefined;
-    }
-    const time1 = new Date(isoTime1);
-    const time2 = new Date(isoTime2);
-    const diff = time2.getTime() - time1.getTime();
-    const minutes = Math.floor(diff / 60_000);
-    return minutes;
   }
 }
